Move verify locks functionality to task
